update all deps and add agenda specific linting, etc. files
