+++ package.json
@@ -41,8 +41,8 @@
     "ava": "^0.25.0",
     "babel-cli": "^6.11.4",
     "babel-core": "^6.13.2",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.4.1",
+    "babel-eslint": "^8.2.1",
+    "babel-loader": "^7.1.2",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-syntax-async-functions": "^6.13.0",
     "babel-plugin-syntax-object-rest-spread": "^6.13.0",
@@ -59,23 +59,43 @@
     "babel-plugin-transform-strict-mode": "^6.11.3",
     "babel-preset-es2015": "^6.6.0",
     "babel-register": "^6.7.2",
-    "eslint": "^3.3.1",
+    "eslint": "^4.16.0",
     "eslint-loader": "^1.5.0",
-    "eslint-plugin-babel": "^3.3.0",
+    "eslint-plugin-babel": "^4.1.2",
     "webpack": "^3.5.5",
     "xo": "^0.18.2"
   },
   "dependencies": {
-    "agenda": "^0.9.0",
+    "agenda": "^1.0.3",
     "babel-runtime": "^6.18.0",
     "commander": "^2.9.0",
     "koa": "^2.0.0-alpha.7",
-    "koa-bodyparser": "^2.3.0",
-    "koa-logger": "^1.3.0",
+    "koa-bodyparser": "^4.2.0",
+    "koa-logger": "^3.1.0",
     "koa-router": "^7.0.1",
     "pythonic": "0.0.6",
-    "regenerator-runtime": "^0.10.0",
+    "regenerator-runtime": "^0.11.1",
     "request": "^2.79.0",
     "request-promise": "^4.1.1"
+},
+"xo": {
+    "space": 2,
+    "rules": {
+      "space-before-function-paren": [
+        "error",
+        "never"
+      ],
+      "max-params": [
+        "error",
+        5
+      ],
+      "max-nested-callbacks": [
+        "error",
+        5
+      ]
+    },
+    "envs": [
+      "node"
+    ]
   }
 }
