+++ src/index.js
@@ -1,175 +1,112 @@
-import querystring from 'querystring';
 import Agenda from 'agenda';
 import Koa from 'koa';
 import logger from 'koa-logger';
 import Router from 'koa-router';
 import bodyParser from 'koa-bodyparser';
-import {keyValues} from 'pythonic';
-import rp from 'request-promise';
 import settings from './settings';
+import {defineJob, newCheck} from './utils';
 
 const app = new Koa();
 const router = new Router();
 app.use(logger());
 app.use(bodyParser({
-	onerror(error, ctx) {
-		ctx.throw(`cannot parse request body, ${JSON.stringify(error)}`, 400);
-	}
+  onerror(error, ctx) {
+    ctx.throw(`cannot parse request body, ${JSON.stringify(error)}`, 400);
+  }
 }));
 app.use(router.routes());
 
 const agenda = new Agenda({
-	db: {
-		address: settings.agendaMongoUrl,
-		collection: settings.collection
-	}
+  db: {
+    address: settings.agendaMongoUrl,
+    collection: settings.collection
+  }
 });
 const mongoDb = () => agenda._mdb;
 let jobs;
 
-const defineJob = ({name, url, method, callback} = {}) => {
-	agenda.define(name, (job, done) => {
-		const data = job.attrs.data;
-		let uri = url;
-		for (const [key, value] of keyValues(data.params)) {
-			uri = uri.replace(`:${key}`, value);
-		}
-		const query = querystring.stringify(data.query);
-		if (query !== '') {
-			uri += `?${query}`;
-		}
-		Promise.race([
-			new Promise((resolve, reject) => setTimeout(() =>
-                reject(new Error('TimeOutError')), settings.timeout)),
-			rp({
-				method: method || 'POST',
-				uri,
-				body: data.body,
-				headers: data.headers || {},
-				json: true
-			})
-		])
-            .catch(err => {
-	job.fail(err.message);
-	return {error: err.message};
-})
-            .then(result => {
-	if (callback) {
-		return rp({
-			method: callback.method || 'POST',
-			uri: callback.url,
-			headers: callback.headers || {},
-			body: {data: data.body, response: result},
-			json: true
-		});
-	}
-})
-            .catch(err => job.fail(`failure in callback: ${err.message}`))
-            .then(done);
-	});
-
-	jobs.count({name}, (error, count) => {
-		if (error) {
-			return console.dir(error);
-		}
-		if (count < 1) {
-			jobs.insert({name, url, method, callback});
-		} else {
-			jobs.update({name}, {$set: {url, method, callback}});
-		}
-	});
-
-	return 'job defined';
-};
-
 agenda.on('ready', () => {
-	jobs = mongoDb().collection(settings.definitions);
-	jobs.find().each((error, job) => {
-		if (!job) {
-			return;
-		}
-		defineJob(job);
-	});
+  jobs = mongoDb().collection(settings.definitions);
+  jobs.find().each((error, job) => {
+    if (!job) {
+      return;
+    }
+    defineJob(agenda, jobs, job);
+  });
 
-	agenda.start();
+  agenda.start();
 });
 
-const newCheck = body => {
-	if (!body.name || !body.url) {
-		throw new Error('expected request body to match {name, url}');
-	}
-	return body;
-};
-
 router.post('/api/new', async (ctx, next) => {
-	ctx.body = await Promise.resolve(ctx.request.body)
-        .then(newCheck)
-        .then(defineJob)
-        .catch(err => {
-			ctx.status = 400;
-			return err.message;
-		});
-	await next();
+  ctx.body = await Promise.resolve(ctx.request.body)
+    .then(newCheck)
+    .then(defineJob)
+    .catch(err => {
+      ctx.status = 400;
+      return err.message;
+    });
+  await next();
 });
 
 const scheduleCheckAndFillMissing = body => {
-	if (!body.name || !body.human_interval) {
-		throw new Error('expected request body to match {name, human_interval}');
-	}
-	if (!body.data) {
-		body.data = {};
-	}
-	if (!body.data.body) {
-		body.data.body = {};
-	}
-	if (!body.data.params) {
-		body.data.params = {};
-	}
-	if (!body.data.query) {
-		body.data.query = {};
-	}
-	return body;
+  if (!body.name || !body.human_interval) {
+    throw new Error('expected request body to match {name, human_interval}');
+  }
+  if (!body.data) {
+    body.data = {};
+  }
+  if (!body.data.body) {
+    body.data.body = {};
+  }
+  if (!body.data.params) {
+    body.data.params = {};
+  }
+  if (!body.data.query) {
+    body.data.query = {};
+  }
+  return body;
 };
 
 const schedulePromise = (ctx, scheduleOrEvery) => Promise.resolve(ctx.request.body)
     .then(scheduleCheckAndFillMissing)
     .then(body => {
-	agenda[scheduleOrEvery].apply(agenda, [body.human_interval, body.name, body.data]);
-	return `job scheduled${scheduleOrEvery === 'every' ? ' for repetition' : ''}`;
-})
-    .catch(error => {
-	ctx.status = 400;
-	return error.message;
-});
+      // eslint-disable-next-line no-useless-call
+      agenda[scheduleOrEvery].apply(agenda, [body.human_interval, body.name, body.data]);
+      return `job scheduled${scheduleOrEvery === 'every' ? ' for repetition' : ''}`;
+    })
+    .catch(err => {
+      ctx.status = 400;
+      return err.message;
+    });
 
 router.post('/api/schedule', async (ctx, next) => {
-	ctx.body = await schedulePromise(ctx, 'schedule');
-	await next();
+  ctx.body = await schedulePromise(ctx, 'schedule');
+  await next();
 });
 
 router.post('/api/every', async (ctx, next) => {
-	ctx.body = await schedulePromise(ctx, 'every');
-	await next();
+  ctx.body = await schedulePromise(ctx, 'every');
+  await next();
 });
 
 router.post('/api/cancel', async (ctx, next) => {
-	agenda.cancel(ctx.request.body, (error, numRemoved) => {
-		if (error) {
-			ctx.status = 400;
-			ctx.body = error.message;
-			return next();
-		}
-		console.log(`${numRemoved} jobs removed`);
-		ctx.body = numRemoved;
-	});
+  agenda.cancel(ctx.request.body, (error, numRemoved) => {
+    if (error) {
+      ctx.status = 400;
+      ctx.body = error.message;
+      return next();
+    }
+    console.log(`${numRemoved} jobs removed`);
+    ctx.body = numRemoved;
+  });
 });
 
 const graceful = () => {
-	console.log('\nShutting down gracefully...');
-	agenda.stop(() => {
-        // eslint-disable-line unicorn/no-process-exit
-		process.exit(0);
-	});
+  console.log('\nShutting down gracefully...');
+  agenda.stop(() => {
+    // eslint-disable-next-line unicorn/no-process-exit
+    process.exit(0);
+  });
 };
 
 process.on('SIGTERM', graceful);
