+++ src/utils/new-check.js
@@ -1,0 +1,8 @@
+const newCheck = body => {
+  if (!body.name || !body.url) {
+    throw new Error('expected request body to match {name, url}');
+  }
+  return body;
+};
+
+export default newCheck;
